states/buffs tracking
* added macros +State, -State and +Time. These are used to manage states
on tokens. These new macros allow you to attach a "state" from the list
of states in "Tokens" group in the campaign and apply it to a token with
a spell name and duration. Time is automatically advanced for combat
during "Next" macro for rounds but you can also manually managed with
+Time if the buffs last outside of that encounter.